Sign up not navigate to login screen
